make endpoints work and test them with insomnia

+++ src/main/java/br/senai/lab365/Petshop/models/Pet.java
@@ -13,11 +13,11 @@ public class Pet {
     private String birthDate;
     private Map<LocalDate, Double> weightHistory;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
